GPL of ASUS RT-N16 for firmware 3.0.0.4.380.7378

+++ release/src/router/www/QoS_EZQoS.asp
@@ -245,7 +245,7 @@
 <% wanlink(); %>
 <% first_wanlink(); %>
 var wans_dualwan_orig = '<% nvram_get("wans_dualwan"); %>';
-var wans_flag = (wans_dualwan_orig.search("none") == -1) ? 1:0;
+var wans_flag = (wans_dualwan_orig.search("none") != -1 || !parent.dualWAN_support) ? 0 : 1;
 var dsllink_statusstr = "";
 if(wans_flag == 1)	//dual_wan enabled
 	dsllink_statusstr = first_wanlink_statusstr();
@@ -278,7 +278,8 @@ function initial(){
 	}
 	
 	if(document.form.qos_enable_orig.value == 1){
-		if(document.form.qos_type.value == 2){		//Bandwidth Limiter
+		if(document.form.qos_type.value == 2){		
+			// Bandwidth Limiter
 			document.getElementById('upload_tr').style.display = "none";
 			document.getElementById('download_tr').style.display = "none";
 			genMain_table();
@@ -289,7 +290,8 @@ function initial(){
 			
 			showDropdownClientList('setClientIP', 'name>mac', 'all', 'ClientList_Block_PC', 'pull_arrow', 'all');
 		}
-		else if(document.form.qos_type.value == 1){		//Adaptive QoS
+		else if(document.form.qos_type.value == 1){		
+			// Adaptive QoS
 			if(document.getElementById("auto").checked){
 				document.getElementById('upload_tr').style.display = "";
 				document.getElementById('download_tr').style.display = "";
@@ -299,9 +301,10 @@ function initial(){
 				document.getElementById('download_tr').style.display = "none";
 			}
 		}
-		else{
-			document.getElementById('upload_tr').style.display = "none";
-			document.getElementById('download_tr').style.display = "none";
+		else{ 
+			// Traditional QoS
+			document.getElementById('upload_tr').style.display = "";
+			document.getElementById('download_tr').style.display = "";
 		}
 
 		document.getElementById('qos_type_tr').style.display = "";
@@ -393,50 +396,58 @@ function switchPage(page){
 		return false;
 }
 
-function submitQoS(){
+function validForm(){
+
 	if(document.form.qos_enable.value == 0 && document.form.qos_enable_orig.value == 0){
 		return false;
 	}
 
 	if(document.form.qos_enable.value == 1){
-		if(document.form.qos_type.value != 2){
-			if(document.form.obw.value.length == 0){	//To check field is empty
+		if(document.form.qos_type.value != 2){	//not Bandwidth Limiter
+
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.obw.value.length == 0){	//To check field is empty
 				alert("<#JS_fieldblank#>");
 				document.form.obw.focus();
 				document.form.obw.select();
-				return;
+				return false;
 			}
-			else if(document.form.qos_type.value == 0 && document.form.obw.value == 0){		// To check field is 0 && Traditional QoS
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.obw.value == 0){		// To check field is 0 && Traditional QoS
 				alert("Upload Bandwidth can not be 0");	/* untranslated */
 				document.form.obw.focus();
 				document.form.obw.select();
-				return;
+				return false;
 			
 			}
-			else if(document.form.obw.value.split(".").length > 2){		//To check more than two point symbol
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.obw.value.split(".").length > 2){		//To check more than two point symbol
 				alert("The format of field of upload bandwidth is invalid"); /* untranslated */
 				document.form.obw.focus();
 				document.form.obw.select();
-				return;	
+				return false;
 			}
-			
-			if(document.form.ibw.value.length == 0){
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && !validator.range(document.form.obw, 1, 9999999999)){
+				return false;
+			}
+					
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.ibw.value.length == 0){
 				alert("<#JS_fieldblank#>");
 				document.form.ibw.focus();
 				document.form.ibw.select();
-				return;
+				return false;
 			}
-			else if(document.form.qos_type.value == 0 && document.form.ibw.value == 0){		// To check field is 0 && Traditional QoS
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.ibw.value == 0){		// To check field is 0 && Traditional QoS
 				alert("Download Bandwidth can not be 0");	/* untranslated */
 				document.form.ibw.focus();
 				document.form.ibw.select();
-				return;
+				return false;
 			}
-			else if(document.form.ibw.value.split(".").length > 2){
+			if(((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && document.form.ibw.value.split(".").length > 2){
 				alert("The format of field of download bandwidth is invalid");	/* untranslated */
 				document.form.ibw.focus();
 				document.form.ibw.select();
-				return;	
+				return false;
+			}
+			if( ((document.form.qos_type.value == 1 && document.form.bw_setting_name[1].checked == true ) || document.form.qos_type.value == 0) && !validator.range(document.form.ibw, 1, 9999999999)){
+				return false;
 			}
 			
 			if(document.form.qos_type.value == 1 && document.getElementById('auto').checked){
@@ -462,7 +473,7 @@ function submitQoS(){
 						document.form.bwdpi_app_rulelist.value = bwdpi_app_rulelist;	
 				}
 				else{
-					alert("You have not selected QoS priority mode.");	
+					alert("You have not selected QoS priority mode.");		//untranslated
 					return false;
 				}							
 			}
@@ -475,38 +486,42 @@ function submitQoS(){
 			
 			document.form.qos_bw_rulelist.value = qos_bw_rulelist;	
 		}
-	}	
-	
-	if(ctf_disable == 1){
-		document.form.action_script.value = "restart_qos;restart_firewall";
 	}
-	else{
-		if(ctf_fa_mode == "2"){
-			FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
+
+	return true;
+}
+
+function submitQoS(){
+
+	if(validForm()){
+
+		if(document.form.qos_enable.value == 1 && document.form.qos_type.value == 1 && document.form.TM_EULA.value == 0){
+			show_tm_eula();
 		}
 		else{
-			if(document.form.qos_type.value == 0)
-				FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
-			else{				
+			if(ctf_disable == 1){
 				document.form.action_script.value = "restart_qos;restart_firewall";
-			}					
-		}
-	}
-	
-	/*if(document.form.qos_type.value == 2){
-		if((document.form.qos_enable.value != document.form.qos_enable_orig.value) && document.form.qos_enable.value == 0){
-			document.form.action_script.value = "restart_qos;restart_firewall";
+			}
+			else{
+				if(ctf_fa_mode == "2"){
+					FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
+				}
+				else{
+					if(document.form.qos_type.value == 0){
+						FormActions("start_apply.htm", "apply", "reboot", "<% get_default_reboot_time(); %>");
+					}
+					else{
+						document.form.action_script.value = "restart_qos;restart_firewall";
+					}
+				}
+			}
+
 			showLoading();
-			document.form.submit();		
-		}
-		else{
-			location.href = "Bandwidth_Limiter.asp";
+			document.form.submit();
 		}
 	}
-	else{*/
-		showLoading();
-		document.form.submit();		
-	//}
+	else
+		return;
 }
 
 function change_qos_type(value){
@@ -817,14 +832,19 @@ function addRow_main(obj, length){
 	if(document.form.PC_devicename.value == "" || document.getElementById("download_rate").value == "" || document.getElementById("upload_rate").value == ""){
 		return true;
 	}
-	
+
 	var enable_checkbox = $(obj.parentNode).siblings()[0].children[0];
 	var invalid_char = "";
 	var qos_bw_rulelist_row =  qos_bw_rulelist.split("<");
 	var max_priority = 0;
 	if(qos_bw_rulelist != "")
-		max_priority = qos_bw_rulelist_row.length;	
-	
+		max_priority = qos_bw_rulelist_row.length;
+
+	if(qos_bw_rulelist_row.length >= length){
+		alert("<#JS_itemlimit1#> " + length + " <#JS_itemlimit2#>");
+		return false;   
+	}
+
 	if(!validator.string(document.form.PC_devicename))
 		return false;
 	
@@ -945,7 +965,7 @@ function genMain_table(){
 	code += '</td>';
 	code += '<td style="border-bottom:2px solid #000;text-align:right;"><input type="text" id="download_rate" class="input_6_table" maxlength="6" onkeypress="return bandwidth_code(this, event);" onkeyup="check_field();"><span style="margin: 0 5px;color:#FFF;">Mb/s</span></td>';
 	code += '<td style="border-bottom:2px solid #000;text-align:right;"><input type="text" id="upload_rate" class="input_6_table" maxlength="6" onkeypress="return bandwidth_code(this, event);" onkeyup="check_field();"><span style="margin: 0 5px;color:#FFF;">Mb/s</span></td>';
-	code += '<td style="border-bottom:2px solid #000;"><div id="add_delete" class="add_disable" style="width:23px;height:23px;margin:0 auto" onclick="addRow_main(this, 32)"></div></td>';
+	code += '<td style="border-bottom:2px solid #000;"><div id="add_delete" class="add_disable" style="margin:0 auto" onclick="addRow_main(this, 32)"></div></td>';
 	code += '</tr>';
 	
 	if(qos_bw_rulelist == ""){
@@ -983,7 +1003,7 @@ function genMain_table(){
 			
 			code += '<td style="text-align:center;">'+qos_bw_rulelist_col[2]/1024+' Mb/s</td>';
 			code += '<td style="text-align:center;">'+qos_bw_rulelist_col[3]/1024+' Mb/s</td>';
-			code += '<td><div class="remove" style="width:23px;height:23px;margin:0 auto" onclick="deleteRow_main(this);"></td>';
+			code += '<td><div class="remove" style="margin:0 auto" onclick="deleteRow_main(this);"></td>';
 			code += '</tr>';
 
 		}
@@ -1033,35 +1053,17 @@ function device_filter(obj){
 		for(var i = 0; i < clientList.length; i += 1) {
 			var clientObj = clientList[clientList[i]];
 			var clientName = (clientObj.nickName == "") ? clientObj.name : clientObj.nickName;
-			if(clientList[i].indexOf(obj.value) == -1 && clientName.indexOf(obj.value) == -1)
+			if(clientList[i].toLowerCase().indexOf(obj.value.toLowerCase()) == -1 && clientName.toLowerCase().indexOf(obj.value.toLowerCase()) == -1)
 				continue;
 			
-			code += '<a title=' + clientList[i] + '><div style="height:auto;" onclick="setClientIP(\'' + clientName + '\', \'' + clientObj.mac + '\');"><strong>' + clientName + '</strong> ';
-			code += ' </div></a>';
+			code += '<div><a title=' + clientList[i] + '><div style="height:auto;" onclick="setClientIP(\'' + clientName + '\', \'' + clientObj.mac + '\');"><strong>' + clientName + '</strong> ';
+			code += ' </div><!--[if lte IE 6.5]><iframe class="hackiframe2"></iframe><![endif]--></a></div>';
 		}		
 		
-		document.getElementById("ClientList_Block_PC").innerHTML = code;		
-	}
-}
+		document.getElementById("ClientList_Block_PC").innerHTML = code;
 
-var qos_enable_ori = "<% nvram_get("qos_enable"); %>";					
-function applyRule(){
-	var qos_bw_rulelist_row = "";
-	if(document.form.PC_devicename.value != ""){
-		alert("You must press add icon to add a new rule first.");	//untranslated
-		return false;
+		if(document.getElementById("ClientList_Block_PC").childNodes.length == "0")	hideClients_Block();
 	}
-	
-	document.form.qos_bw_rulelist.value = qos_bw_rulelist;
-	document.form.qos_enable.value = 1;
-	document.form.qos_type.value = 2;
-	if(ctf_disable == 0 && ctf_fa_mode == 2){
-		document.form.action_script.value = "reboot";
-		document.form.action_wait.value = "<% nvram_get("reboot_time"); %>";
-	}
-
-	showLoading();	
-	document.form.submit();
 }
 
 function enable_check(obj){
@@ -1111,11 +1113,43 @@ function check_field(){
 		$("#add_delete").removeClass("add_enable").addClass("add_disable");
 	}
 }
+
+function show_tm_eula(){
+	if(document.form.preferred_lang.value == "JP"){
+		$.get("JP_tm_eula.htm", function(data){
+			document.getElementById('agreement_panel').innerHTML= data;
+
+		});             
+	}
+	else{
+		$.get("tm_eula.htm", function(data){
+			document.getElementById('agreement_panel').innerHTML= data;
+
+		});
+	}
+	dr_advise();
+	cal_panel_block("agreement_panel", 0.25);
+	$("#agreement_panel").fadeIn(300);
+}
+
+function eula_confirm(){
+	document.form.TM_EULA.value = 1;
+	submitQoS();
+}
+
+function cancel(){
+	refreshpage();
+}
 </script>
 </head>	
 <body onload="initial();" id="body_id" onunload="unload_body();" onClick="">	
 <div id="TopBanner"></div>
 <div id="Loading" class="popup_bg"></div>
+<div id="agreement_panel" class="panel_folder" style="margin-top: -100px;"></div>
+<div id="hiddenMask" class="popup_bg" style="z-index:999;">
+	<table cellpadding="5" cellspacing="0" id="dr_sweet_advise" class="dr_sweet_advise" align="center"></table>
+	<!--[if lte IE 6.5.]><script>alert("<#ALERT_TO_CHANGE_BROWSER#>");</script><![endif]-->
+</div>
 
 <iframe name="hidden_frame" id="hidden_frame" width="0" height="0" frameborder="0"></iframe>
 <table id="main_table" class="content" align="center" cellpadding="0" cellspacing="0">
@@ -1199,9 +1233,11 @@ function check_field(){
 			<input type="hidden" name="action_script" value="">
 			<input type="hidden" name="action_wait" value="5">
 			<input type="hidden" name="flag" value="">
+			<input type="hidden" name="TM_EULA" value="<% nvram_get("TM_EULA"); %>">
 			<input type="hidden" name="qos_enable" value="<% nvram_get("qos_enable"); %>">
 			<input type="hidden" name="qos_enable_orig" value="<% nvram_get("qos_enable"); %>">
 			<input type="hidden" name="qos_type_orig" value="<% nvram_get("qos_type"); %>">
+			<input type="hidden" name="qos_type" value="<% nvram_get("qos_type"); %>">
 			<input type="hidden" name="qos_obw" value="<% nvram_get("qos_obw"); %>" disabled>
 			<input type="hidden" name="qos_ibw" value="<% nvram_get("qos_ibw"); %>" disabled>
 			<input type="hidden" name="bwdpi_app_rulelist" value="<% nvram_get("bwdpi_app_rulelist"); %>" disabled>
@@ -1252,7 +1288,7 @@ function check_field(){
 														<#EzQoS_desc_note#>														
 													</div>
 													<div class="formfontdesc">
-														<a id="faq" href="http://www.asus.com/us/support/FAQ/1008718/" target="_blank" style="text-decoration:underline;">QoS FAQ</a>
+														<a id="faq" href="http://www.asus.com/support/FAQ/1008718/" target="_blank" style="text-decoration:underline;">QoS FAQ</a>
 													</div>
 												</td>
 											</tr>
@@ -1315,9 +1351,9 @@ function check_field(){
 										<tr id="qos_type_tr" style="display:none">
 											<th><#QoS_Type#></th>
 											<td colspan="2">
-												<input id="int_type" name="qos_type" value="1" onClick="change_qos_type(this.value);" style="display:none;" type="radio" <% nvram_match("qos_type", "1","checked"); %>><a id="int_type_link" class="hintstyle" style="display:none;" href="javascript:void(0);" onClick="openHint(20, 6);"><label for="int_type"><#Adaptive_QoS#></label></a>
-												<input id="trad_type" name="qos_type" value="0" onClick="change_qos_type(this.value);" type="radio" <% nvram_match("qos_type", "0","checked"); %>><a class="hintstyle" href="javascript:void(0);" onClick="openHint(20, 7);"><label for="trad_type"><#EzQoS_type_traditional#></label></a>
-												<input id="bw_limit_type" name="qos_type" value="2" onClick="change_qos_type(this.value);" type="radio" <% nvram_match("qos_type", "2","checked"); %>><a class="hintstyle" href="javascript:void(0);" onClick="openHint(20, 8)"><label for="bw_limit_type"><#Bandwidth_Limiter#></label></a>
+												<input id="int_type" name="qos_type_radio" value="1" onClick="change_qos_type(this.value);" style="display:none;" type="radio" <% nvram_match("qos_type", "1","checked"); %>><a id="int_type_link" class="hintstyle" style="display:none;" href="javascript:void(0);" onClick="openHint(20, 6);"><label for="int_type"><#Adaptive_QoS#></label></a>
+												<input id="trad_type" name="qos_type_radio" value="0" onClick="change_qos_type(this.value);" type="radio" <% nvram_match("qos_type", "0","checked"); %>><a class="hintstyle" href="javascript:void(0);" onClick="openHint(20, 7);"><label for="trad_type"><#EzQoS_type_traditional#></label></a>
+												<input id="bw_limit_type" name="qos_type_radio" value="2" onClick="change_qos_type(this.value);" type="radio" <% nvram_match("qos_type", "2","checked"); %>><a class="hintstyle" href="javascript:void(0);" onClick="openHint(20, 8)"><label for="bw_limit_type"><#Bandwidth_Limiter#></label></a>
 											</td>
 										</tr>
 										<tr id="bandwidth_setting_tr" style="display:none">
@@ -1410,7 +1446,7 @@ function check_field(){
 						<table width="100%">
 							<tr>
 								<td height="50" >
-									<div style=" *width:136px;margin-left:300px;" class="titlebtn" align="center" onClick="submitQoS()"><span><#CTL_apply#></span></div>
+									<div style=" *width:136px;margin-left:300px;" class="titlebtn" align="center" onClick="submitQoS();"><span><#CTL_apply#></span></div>
 								</td>
 							</tr>					
 						</table>
